Refactor: generalizar la clase Product para mejorar la reutilización y flexibilidad del código
